Moving utils to a seperate utils file
